feat: version 3

+++ main.py
@@ -3,9 +3,8 @@ import logging
 import os
 import signal
 import threading
-import time
 from datetime import datetime
-from multiprocessing import Event, Process, Queue
+from queue import Empty, Queue
 
 import cv2
 import numpy as np
@@ -17,17 +16,22 @@ logging.basicConfig(level=logging.INFO, format="%(asctime)s - %(levelname)s - %(
 
 
 class Config:
-    ENABLE_VISUALIZATION = False  # Controla si se muestra la ventana de video (True/False)
-    DETECTION_LIST = ["person"]  # Lista de objetos a detectar
-    CONFIDENCE_THRESHOLD = 0.15  # Confianza mínima para detectar objetos
-    OUTPUT_FOLDER = "output"  # Carpeta de salida para imágenes y CSV
+    ENABLE_VISUALIZATION = False     # Controla si se muestra la ventana de video (True/False)
+    DETECTION_LIST = ["person"]      # Lista de objetos a detectar
+    CONFIDENCE_THRESHOLD = 0.15      # Confianza mínima para detectar objetos
+    OUTPUT_FOLDER = "output"         # Carpeta de salida para imágenes y CSV
     CSV_FILENAME = "detections.csv"  # Nombre del archivo CSV para almacenar resultados
-    MIN_MOTION_AREA = 5000  # Área mínima de movimiento para considerar que hay cambio
-    RESIZE_WIDTH = 640  # Ancho al que se redimensionará la imagen para detección de movimiento
-    CAMERA_RESOLUTION = (640, 360)  # Resolución de la cámara (ancho, alto)
-    MODEL_NAME = "yolov8x.pt"  # Modelo YOLOv8 a utilizar (nano para mayor velocidad)
-    CSV_BUFFER_SIZE = 2  # Número de entradas antes de escribir en el CSV
-    MOTION_DETECTION_COOLDOWN = 5  # Segundos de espera después de detectar movimiento
+    MIN_MOTION_AREA = 5000           # Área mínima de movimiento para considerar que hay cambio
+    RESIZE_WIDTH = 640               # Ancho al que se redimensionará la imagen para detección de movimiento
+    CAMERA_RESOLUTION = (640, 360)   # Resolución de la cámara (ancho, alto)
+    MODEL_NAME = "yolov8x.pt"        # Modelo YOLOv8 a utilizar
+    CSV_BUFFER_SIZE = 2              # Número de entradas antes de escribir en el CSV
+    MOTION_DETECTION_COOLDOWN = 5    # Segundos de espera después de detectar movimiento
+    BACKGROUND_SUBTRACTOR = {        # Configuración para createBackgroundSubtractorKNN
+        "history": 500,
+        "dist2Threshold": 400.0,
+        "detectShadows": False,
+    }
 
 
 class ObjectDetector:
@@ -42,7 +46,7 @@ class ObjectDetector:
 
     def get_classes_ids(self):
         if not self.config.DETECTION_LIST:
-            return None  # Detectar todas las clases si la lista está vacía
+            return None
         classes_ids = []
         for object_name in self.config.DETECTION_LIST:
             if object_name in self.model.names.values():
@@ -93,14 +97,20 @@ class FrameCapture(threading.Thread):
     """Hilo para capturar frames de la cámara y detectar movimiento"""
 
     def __init__(self, config, frame_queue, motion_event, stop_event):
-        threading.Thread.__init__(self)
+        super().__init__()
         self.config = config
         self.frame_queue = frame_queue
         self.motion_event = motion_event
         self.stop_event = stop_event
         self.picam2 = Picamera2()
         self.configure_camera()
-        logging.info("Cámara inicializada y configurada.")
+        # Configurar el sustractor de fondo KNN para detección de movimiento
+        self.backSub = cv2.createBackgroundSubtractorKNN(
+            history=self.config.BACKGROUND_SUBTRACTOR["history"],
+            dist2Threshold=self.config.BACKGROUND_SUBTRACTOR["dist2Threshold"],
+            detectShadows=self.config.BACKGROUND_SUBTRACTOR["detectShadows"],
+        )
+        logging.info("Cámara y sustractor de fondo inicializados y configurados.")
 
     def configure_camera(self):
         # Configurar la cámara con la resolución especificada
@@ -109,72 +119,65 @@ class FrameCapture(threading.Thread):
         self.picam2.start()
 
     def run(self):
-        prev_frame = None
         last_motion_time = None
 
         while not self.stop_event.is_set():
-            # Capturar un frame de la cámara
-            frame = self.picam2.capture_array()
+            try:
+                # Capturar un frame de la cámara
+                frame = self.picam2.capture_array()
 
-            # Convertir el frame a escala de grises para detección de movimiento
-            gray = cv2.cvtColor(frame, cv2.COLOR_BGR2GRAY)
-            gray = cv2.resize(
-                gray, (self.config.RESIZE_WIDTH, int(gray.shape[0] * self.config.RESIZE_WIDTH / gray.shape[1]))
-            )
-            gray = cv2.GaussianBlur(gray, (21, 21), 0)
+                # Aplicar el sustractor de fondo para detectar movimiento
+                fg_mask = self.backSub.apply(frame)
 
-            if prev_frame is None:
-                prev_frame = gray
-                continue
+                # Operaciones morfológicas para reducir el ruido
+                kernel = cv2.getStructuringElement(cv2.MORPH_ELLIPSE, (3, 3))
+                fg_mask = cv2.morphologyEx(fg_mask, cv2.MORPH_OPEN, kernel)
 
-            # Calcular la diferencia entre el frame actual y el anterior
-            frame_delta = cv2.absdiff(prev_frame, gray)
-            thresh = cv2.threshold(frame_delta, 25, 255, cv2.THRESH_BINARY)[1]
-            thresh = cv2.dilate(thresh, None, iterations=2)
-            contours, _ = cv2.findContours(thresh.copy(), cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
+                # Encontrar contornos
+                contours, _ = cv2.findContours(fg_mask, cv2.RETR_EXTERNAL, cv2.CHAIN_APPROX_SIMPLE)
 
-            motion_detected = False
-            for contour in contours:
-                if cv2.contourArea(contour) < self.config.MIN_MOTION_AREA:
-                    continue
-                motion_detected = True
-                break
+                motion_detected = False
+                for contour in contours:
+                    if cv2.contourArea(contour) < self.config.MIN_MOTION_AREA:
+                        continue
+                    motion_detected = True
+                    break
 
-            current_time = datetime.now()
+                current_time = datetime.now()
 
-            if motion_detected:
-                if (
-                    last_motion_time is None
-                    or (current_time - last_motion_time).total_seconds() > self.config.MOTION_DETECTION_COOLDOWN
-                ):
-                    logging.info("Movimiento detectado. Iniciando captura.")
-                    self.frame_queue.put(frame)
-                    last_motion_time = current_time
-                    # Señalar que se ha detectado movimiento
-                    self.motion_event.set()
+                if motion_detected:
+                    if (
+                        last_motion_time is None
+                        or (current_time - last_motion_time).total_seconds() > self.config.MOTION_DETECTION_COOLDOWN
+                    ):
+                        logging.info("Movimiento detectado. Encolando frame para detección.")
+                        self.frame_queue.put(frame)
+                        last_motion_time = current_time
+                        # Señalar que se ha detectado movimiento
+                        self.motion_event.set()
                 else:
-                    logging.debug("Movimiento detectado, pero en periodo de enfriamiento.")
-            else:
-                # Si ha pasado el periodo de enfriamiento, resetear el evento
-                if (
-                    last_motion_time
-                    and (current_time - last_motion_time).total_seconds() > self.config.MOTION_DETECTION_COOLDOWN
-                ):
-                    self.motion_event.clear()
-                    last_motion_time = None
+                    # Si ha pasado el periodo de enfriamiento, resetear el evento
+                    if (
+                        last_motion_time
+                        and (current_time - last_motion_time).total_seconds() > self.config.MOTION_DETECTION_COOLDOWN
+                    ):
+                        self.motion_event.clear()
+                        last_motion_time = None
 
-            prev_frame = gray
+            except Exception as e:
+                logging.error(f"Error en FrameCapture: {e}")
 
     def stop(self):
         self.stop_event.set()
         self.picam2.stop()
-        logging.info("Cámara liberada.")
+        logging.info("Cámara liberada y FrameCapture detenido.")
 
 
-class DetectorProcess:
-    """Proceso para la detección de objetos"""
+class DetectorThread(threading.Thread):
+    """Hilo para la detección de objetos"""
 
     def __init__(self, config, frame_queue, motion_event, stop_event):
+        super().__init__()
         self.config = config
         self.frame_queue = frame_queue
         self.motion_event = motion_event
@@ -187,48 +190,47 @@ class DetectorProcess:
     def run(self):
         try:
             while not self.stop_event.is_set():
-                if self.motion_event.is_set():
-                    # Procesar frames mientras se ha detectado movimiento
-                    if not self.frame_queue.empty():
-                        frame = self.frame_queue.get()
+                try:
+                    # Esperar hasta que haya un frame disponible o se indique detenerse
+                    frame = self.frame_queue.get(timeout=1)
+                except Empty:
+                    continue
 
-                        # Asegurar que el frame tiene 3 canales
-                        if frame.shape[2] == 4:
-                            frame = cv2.cvtColor(frame, cv2.COLOR_RGBA2BGR)
+                # Asegurar que el frame tiene 3 canales
+                if len(frame.shape) == 3 and frame.shape[2] == 4:
+                    frame = cv2.cvtColor(frame, cv2.COLOR_RGBA2BGR)
 
-                        # Realizar detección
-                        results = self.detector.detect_objects(frame)
+                # Realizar detección
+                results = self.detector.detect_objects(frame)
 
-                        # Filtrar detecciones
-                        detected_objects, confidences, coordinates, areas = self.detector.filter_detections(results)
+                # Filtrar detecciones
+                detected_objects, confidences, coordinates, areas = self.detector.filter_detections(results)
 
-                        # Anotar frame con detecciones
-                        annotated_frame = results[0].plot()
+                # Anotar frame con detecciones
+                annotated_frame = results[0].plot()
 
-                        # Guardar resultados
-                        self.save_results(frame, annotated_frame, detected_objects, confidences, coordinates, areas)
+                # Guardar resultados
+                self.save_results(annotated_frame, detected_objects, confidences, coordinates, areas)
 
-                        # Mostrar visualización si está habilitada
-                        if self.config.ENABLE_VISUALIZATION:
-                            cv2.imshow("Detección en Tiempo Real", annotated_frame)
-                            cv2.waitKey(1)
+                # Mostrar visualización si está habilitada
+                if self.config.ENABLE_VISUALIZATION:
+                    cv2.imshow("Detección en Tiempo Real", annotated_frame)
+                    if cv2.waitKey(1) & 0xFF == ord("q"):
+                        self.stop_event.set()
 
-                        self.frame_count += 1
-                    else:
-                        time.sleep(0.01)  # Pequeña pausa para evitar sobrecarga de CPU
-                else:
-                    time.sleep(0.1)  # Esperar hasta que se detecte movimiento
-        except KeyboardInterrupt:
-            pass
+                self.frame_count += 1
+
+        except Exception as e:
+            logging.error(f"Error en DetectorThread: {e}")
         finally:
             if self.config.ENABLE_VISUALIZATION:
                 cv2.destroyAllWindows()
             # Escribir cualquier resultado pendiente en el CSV
             if self.csv_buffer:
                 self.write_csv_buffer()
-            logging.info("Proceso de detección finalizado.")
+            logging.info("DetectorThread finalizado.")
 
-    def save_results(self, frame, annotated_frame, detected_objects, confidences, coordinates, areas):
+    def save_results(self, annotated_frame, detected_objects, confidences, coordinates, areas):
         # Guardar la imagen anotada con el nombre del frame actual
         logging.info(f"Guardando imagen {self.frame_count:04d}")
         frame_filename = os.path.join(self.config.OUTPUT_FOLDER, f"frame_{self.frame_count:04d}.jpg")
@@ -249,10 +251,14 @@ class DetectorProcess:
             self.write_csv_buffer()
 
     def write_csv_buffer(self):
-        with open(self.csv_filepath, mode="a", newline="") as file:
-            writer = csv.writer(file)
-            writer.writerows(self.csv_buffer)
-        self.csv_buffer = []
+        try:
+            with open(self.csv_filepath, mode="a", newline="") as file:
+                writer = csv.writer(file)
+                writer.writerows(self.csv_buffer)
+            logging.info(f"Escrito {len(self.csv_buffer)} entradas al CSV.")
+            self.csv_buffer = []
+        except Exception as e:
+            logging.error(f"Error al escribir en el CSV: {e}")
 
 
 class DetectionApp:
@@ -261,9 +267,9 @@ class DetectionApp:
     def __init__(self):
         self.config = Config()
         self.frame_queue = Queue()
-        self.motion_event = Event()
-        self.stop_event = Event()
-        self.detector_process = None
+        self.motion_event = threading.Event()
+        self.stop_event = threading.Event()
+        self.detector_thread = None
         self.frame_capture_thread = None
 
         # Crear carpeta de salida si no existe
@@ -276,38 +282,56 @@ class DetectionApp:
         self.init_csv()
 
     def init_csv(self):
-        # Escribir la cabecera del CSV
-        with open(self.csv_filepath, mode="w", newline="") as file:
-            writer = csv.writer(file)
-            writer.writerow(
-                ["Frame", "Fecha", "Hora", "Objetos Detectados", "Confidencias", "Coordenadas", "Tamaño del área"]
-            )
-        logging.info(f"Archivo CSV inicializado: {self.csv_filepath}")
+        # Escribir la cabecera del CSV si el archivo no existe
+        if not os.path.isfile(self.csv_filepath):
+            try:
+                with open(self.csv_filepath, mode="w", newline="") as file:
+                    writer = csv.writer(file)
+                    writer.writerow(
+                        [
+                            "Frame",
+                            "Fecha",
+                            "Hora",
+                            "Objetos Detectados",
+                            "Confidencias",
+                            "Coordenadas",
+                            "Tamaño del área",
+                        ]
+                    )
+                logging.info(f"Archivo CSV inicializado: {self.csv_filepath}")
+            except Exception as e:
+                logging.error(f"Error al inicializar el CSV: {e}")
 
     def start(self):
         # Iniciar hilo de captura de frames
         self.frame_capture_thread = FrameCapture(self.config, self.frame_queue, self.motion_event, self.stop_event)
         self.frame_capture_thread.start()
+        logging.info("FrameCaptureThread iniciado.")
 
-        # Iniciar proceso de detección
-        self.detector_process_instance = DetectorProcess(
-            self.config, self.frame_queue, self.motion_event, self.stop_event
-        )
-        self.detector_process = Process(target=self.detector_process_instance.run)
-        self.detector_process.start()
+        # Iniciar hilo de detección
+        self.detector_thread = DetectorThread(self.config, self.frame_queue, self.motion_event, self.stop_event)
+        self.detector_thread.start()
+        logging.info("DetectorThread iniciado.")
 
-    def signal_handler(self, sig, frame):
+    def stop(self):
         logging.info("Deteniendo la aplicación...")
         self.stop_event.set()
 
         # Detener el hilo de captura de frames
-        self.frame_capture_thread.stop()
-        self.frame_capture_thread.join()
+        if self.frame_capture_thread.is_alive():
+            self.frame_capture_thread.stop()
+            self.frame_capture_thread.join()
+            logging.info("FrameCaptureThread detenido.")
 
-        # Terminar el proceso de detección
-        self.detector_process.join()
+        # Detener el hilo de detección
+        if self.detector_thread.is_alive():
+            self.detector_thread.join()
+            logging.info("DetectorThread detenido.")
 
         logging.info("Aplicación detenida correctamente.")
+
+    def signal_handler(self, sig, frame):
+        self.stop()
 
     def run(self):
         # Manejar señal para Ctrl+C
@@ -315,12 +339,9 @@ class DetectionApp:
         logging.info("Iniciando la aplicación de detección.")
         self.start()
 
-        # Mantener el proceso principal vivo
-        try:
-            while True:
-                time.sleep(1)
-        except KeyboardInterrupt:
-            self.signal_handler(None, None)
+        # Esperar hasta que se indique detenerse
+        self.stop_event.wait()
+        self.stop()
 
 
 if __name__ == "__main__":
